pitch-backend (3)/pitch-backend/pitch-backend/requirements.txt
Flask==3.0.0
flask-pymongo==2.3.0
PyJWT==2.9.0
bcrypt==4.1.3
python-dotenv==1.0.1
Werkzeug==3.0.1
pytz==2024.2
marshmallow==3.22.0
flask-cors==4.0.0
